Creating a PDF outline from the Excel Course Outline
The command to run is >python publishOutlinev2 coursePlan.xlsm  -> which will generate a course outline called coursePlan (or whatever you give as your excel spreadsheet name) using the Example-NWV pdf file as a template.  If the user wishes to walk the directory and use all excel course plans to generate PDF outlines for each, then use >python publishOutlinev2 all.  Note: a textcoursePlan.pdf will be generated for each outline which is the raw text for the watermark.